wip trial of better implimentation of module pattern
